ai fixes, added latex puddles, allowed decimals as airDecreaseSpeed modifier, allowed command blocks to run mod commands

+++ src/main/java/net/zaharenko424/a_changed/transfurSystem/transfurTypes/AbstractTransfurType.java
@@ -138,7 +138,7 @@ public abstract class AbstractTransfurType {
             return this;
         }
 
-        public Properties airReductionModifier(int airReductionModifier){
+        public Properties airReductionModifier(float airReductionModifier){
             this.airReductionModifier = airReductionModifier;
             return this;
         }
